+++ pages/status.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useFormik } from 'formik'
 import * as Yup from 'yup'
-import { CheckStatusRequestBody } from './api/check-status'
+import { CheckStatusRequestBody } from '../lib/StatusTypes'
 import { useCheckStatus } from '../lib/CheckStatusHook'
 import Layout from '../components/Layout'
 import InputField from '../components/InputField'
